% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Creates a grid of points within the contiguous US.}
\usage{
create_grid(
  width = 0.5,
  start_lat = NULL,
  end_lat = NULL,
  start_lon = NULL,
  end_lon = NULL,
  elevation = FALSE
)
}
\arguments{
\item{width}{(optional) Distance in decimal degrees between two neighboring points
on the grid. Default is 0.5.}

\item{start_lat}{(optional) Starting latitude in decimal degrees, ranging 25.1299
to 49.3832. Should be smaller than end_lat if both specified. Default is NULL,
indicating 25.1299 (southernmost point of contiguous US).}

\item{end_lat}{(optional) Ending latitude in decimal degrees, range 25.1299 to
49.3832. Should be larger than start_lat if both specified. Default is NULL,
 indicating 49.3832 (northernmost point of contiguous US).}

\item{start_lon}{(optional) Starting longitude in decimal degrees, range -124.6813
to -67.0074. Should be smaller than end_lon if both specified. Default is NULL,
indicating -124.6813 (westernmost point of contiguous US).}

\item{end_lon}{(optional) Ending longitude in decimal degrees, range -124.6813 to
-67.0074. Should be larger than start_lon if both specified. Default is NULL,
 indicating -67.0074 (easternmost point of contiguous US).}

\item{elevation}{(optional) Boolean indicating whether to include the elevation}
}
\value{
a data frame with the following columns:
\itemize{
    \item \code{latitude} Latitude in decimal degrees of the grid point.
    \item \code{longitude} Longitude in decimal degrees of the grid point.
    \item \code{elevation} Elevation in meters at the grid point (will only be
    provided if user specifies elevation = TRUE).
}
}
\description{
Starting and ending latitudes and longitudes and/or resolution of the grid can be
controlled. Returns a data frame of longitudes, latitudes, and elevation (if
desired) at the grid points within the contiguous USA (all states/regions excluding
 Alaska, Hawaii, and Puerto Rico).
}
\details{
The US map might be slightly different across sources. The map used here comes from
the ggplot2 package.
}
\examples{
grid <- create_grid(width = 0.7)
head(grid)
}
