% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extracts time series data for a weather station.}
\usage{
extract_data(
  dat,
  id,
  var,
  starting_date = "2000-01-01",
  ending_date = "2024-04-30"
)
}
\arguments{
\item{dat}{dataframe with the following columns in sequence: station id
(five-character string), date of observation (Date object), state longitude,
state latitude, temperature-related variables of your interest (can be one or more).}

\item{id}{five-character string representing a weather station you want to extract
data at.}

\item{var}{vector of variable name(s) for the time series. Should be one or more
from the column names of the given dat.}

\item{starting_date}{(optional) Date object for the starting date of the time
series. Default is "2000-01-01".}

\item{ending_date}{(optional) Date object for the ending date of the time series.
Default is "2024-04-30".}
}
\value{
a data frame with the following variables:
\itemize{
    \item \code{station_id} station id of the specified station.
    \item \code{longitude} station longitude.
    \item \code{latitude} station latitude.
    \item \code{lst_date} date of the observation as Date object.
    \item \code{var} time-series data for the specified variables.
}
}
\description{
Extracts time series data of the specified variables for a given weather station
from user-provided data. Starting date and ending date for the time series can be
specified. Only the data available will be returned.
}
\examples{
# extract daily average and minimum temperatures at station "94078" for March, 2024
df <- extract_data(climate_df[,-c(2,3,5)], id = "94078",
                                var = c("t_daily_avg", "t_daily_min"),
                                starting_date = "2024/03/01", ending_date = "2024/03/31")
head(df)
}
