% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_station.R
\docType{data}
\name{station_df}
\alias{station_df}
\title{Weather stations in the U.S.}
\format{
a dataframe with 236 rows and 6 variables:
\describe{
       \item{station_id}{station identifier of the form 'XXXXX'}
       \item{longitude}{station longitude in decimal degrees}
       \item{latitude}{station latitude in decimal degrees}
       \item{state}{state of the station}
       \item{station_name}{station name (e.g. AZ_Tucson_11_W)}
       \item{contiguous}{whether the station is within the contiguous US (i.e., all
                                       states and regions except for Alaska, Hawaii, and Puerto
                                       Rico)}
}
}
\source{
National Centers for Environmental Information (NCEI):
 \url{https://www.ncei.noaa.gov/pub/data/uscrn/products/daily01/}
}
\usage{
station_df
}
\description{
Lists weather stations maintained by the National Centers for Environmental
Information (NCEI) with their ids, names, longitudes, latitudes, states, and
regions.
}
\keyword{datasets}
